requirements.txt
arxiv==2.2.0
faiss_gpu==1.7.2
fastapi==0.116.1
fitz==0.0.1.dev2
fontTools==4.59.0
httpx==0.28.1
huggingface_hub==0.33.1
langchain==0.3.26
langchain_community==0.3.27
langchain_core==0.3.70
langchain_experimental==0.3.4
langchain_openai==0.3.28
langchain_tavily==0.2.10
langgraph==0.5.4
Pillow==11.3.0
pydantic==2.11.7
python-dotenv==1.1.1
Requests==2.32.4
wolframalpha==5.1.3
xmltodict==0.14.2
